requirements.txt
twine>=3.1.1
bokeh>=2.4.2
dask>=2022.02.1
distributed>=2022.02.1
graphviz>=0.19.1
matplotlib>=3.1.2
numba>=0.55.1
numcodecs>=0.9.1
numpy>=1.21.5
pandas>=0.25.2
scipy>=1.5.4
scikit-learn>=0.22.2
toolz>=0.10.0
xarray>=2022.03.0
zarr>=2.11.1
fsspec>=2022.1.0
astropy>=4.3.1
gdown>=3.12.2
ipympl>=0.7.0
pip==21.1.0
nbsphinx==0.8.7
sphinx-autoapi==1.8.0
sphinx-rtd-theme==0.5.0
recommonmark==0.6.0
pypandoc==1.5
Sphinx==3.5.4
pytest>=6.2.4
numba-scipy>=0.3.0
dask-ms>=0.2.7
casatools>=6.3.0.48
casatasks>=6.3.0.48
casadata>=2021.8.23
python-casacore>=3.4.0
cngi-prototype>=1.0.1


dask>=2022.02.1
distributed>=2022.02.1
zarr>=2.11.1
xarray>=2022.03.0
dask-ms>=0.2.7